+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog 
+### Version 1.5.3
+- Add timeout during Instance creation
+- add function alexaResponse( )
+
 ### Version 1.5.2
 - UTF-8 for status buttons
 - allw https for play files